Removing a dependency on Complexion from plugins
Previously the user would use

    tokenizer = new ComplexionSomething();

It was CompexionSomething's job to include Complexion.  Plugins also
couldn't extend others.  Now it looks a little different:

    tokenizer = new Complexion();
    tokenizer.plugin(ComplexionSomething());

ComplexionSomething still has to register itself with the tokenizer, but
does not need to modify prototypes or wrap Complexion in order to get
custom token objects inserted.  It is also easily possible to mock
Complexion and have CompexionSomething attach to the mock instead.  The
plugin can now use configuration like this:

    tokenizer = new Complexion();
    tokenizer.plugin(ComplexionSomething({
        someOption: true
    }));

It felt like the right thing to do in order to untangle dependencies.
Let me know if you have a better idea.